+++ src/test/java/tests/AsyncRestoreTest.java
@@ -2,13 +2,12 @@ package tests;
 
 import exceptions.AsyncRetryException;
 import org.junit.Assert;
-import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.http.HttpHeaders;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-import services.impl.RestoreServiceImpl;
+import services.RestoreFactory;
 import utils.FileUtils;
 import utils.TestContext;
 
@@ -25,15 +24,15 @@ public class AsyncRestoreTest {
     @Test
     public void testHandleAsyncResponse() throws NoSuchMethodException, InvocationTargetException, IllegalAccessException {
         boolean failure = false;
-        RestoreServiceImpl restoreService = new RestoreServiceImpl();
+        RestoreFactory restoreFactory = new RestoreFactory();
         String asyncResponse = FileUtils.getFile(this.getClass(), "restores/async_restore_response.xml");
         HttpHeaders headers = new HttpHeaders();
         headers.add("Retry-After", "30");
-        Method method = restoreService.getClass().getDeclaredMethod("handleAsyncRestoreResponse", String.class, HttpHeaders.class);
+        Method method = restoreFactory.getClass().getDeclaredMethod("handleAsyncRestoreResponse", String.class, HttpHeaders.class);
         method.setAccessible(true);
 
         try {
-            method.invoke(restoreService, asyncResponse, headers);
+            method.invoke(restoreFactory, asyncResponse, headers);
         } catch (InvocationTargetException invocationException) {
             AsyncRetryException e = (AsyncRetryException) invocationException.getTargetException();
             Assert.assertEquals(143, e.getDone());
